+++ CHANGELOG.rst
@@ -23,6 +23,7 @@ Added
     - ``notify_email_irods_request`` user app setting (#1939)
     - Assay app unit tests (#1980)
     - Missing assay plugin ``__init__.py`` files (#2014)
+    - Study plugin override via ISA-Tab comments (#1885)
 
 Changed
 -------